Move jest.d.ts to @types & remove declare global statement
