+++ src/@types/jest.d.ts
@@ -1,0 +1,8 @@
+namespace jest {
+  interface Matchers<R> {
+    toHaveQueryParam(
+      queryParamName: string,
+      expectedQueryParamValue?: unknown
+    ): CustomMatcherResult;
+  }
+}
