--- src/jest.d.ts
@@ -1,10 +1,0 @@
-declare global {
-  namespace jest {
-    interface Matchers<R> {
-      toHaveQueryParam(
-        queryParamName: string,
-        expectedQueryParamValue?: unknown
-      ): CustomMatcherResult;
-    }
-  }
-}
